#25 :truck: domain 모듈로 파일 이동함
