+++ domain/src/main/java/io/github/seoj17/doamain/model/PictureEntity.kt
@@ -1,4 +1,4 @@
-package project.seo.pictureviewer.data.database
+package io.github.seoj17.doamain.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
